pyproject.toml
[project]
name = "clogger"
version = "0.1.0"
description = "OSRS Leagues knowledge base for route planning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
